+++ lib/rake_slack/exceptions.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
 require_relative 'exceptions/delivery_failed'
+require_relative 'exceptions/no_matching_rule'